Servlet
Simple and messy servlet that I made to handle connection setting and member verification.  The values that are returned in HTTP response need some work, but feel free to try and use this.  You must compile this project into a WAR file using maven and then you can deploy to your Tomcat server.